Translation word content once per language not once per book
Requested by content team
